feat: inline scanner in FindPage + delete ScanPage (Find rework complete)
The Scan button now expands an inline ZXing viewfinder (letterbox + green
target reused from ScanPage) instead of pushing a full-screen camera page
(Drew's call — full-screen for a barcode looked ugly). Camera starts on expand
after a runtime permission check, stops on Cancel / tab-leave; on detect it
debounces (3s, like ScanPage), collapses, and pushes ResultPage. Denied
permission falls back to a "type an ISBN" hint.

Deleted ScanPage (+ its DI registration) — the last old light-only page. The
F3 dark-mode stragglers are now fully cleared: every Bookshelf surface is
token-based + dark.

Builds (MAUI Android, 0 warnings).
